Do not allow duplicate criteria in a rubric (#9855)
Do not allow the user to add duplicate criteria to their rubric. The general rule is: is the rubric has not been added, or if it has parameters (thus making it customizable), then it is available to be added. Otherwise, we hide it. If there are no available criteria left to add, we disable the Add Criteria button.