Move Last Active Rubric Name to Browser Local Storage (#9858)
I originally created a metadata table inside the Indexed DB to store this kind of info, but I didn't know about localstorage back then. Since we now have a service to handle that, I wanted to move the last active rubric name into the same area and get rid of that metadata table.