Teacher Tool: Make Splitter Functional (#9887)
This makes it so the splitter can actually be used to resize the windows. I've also made it so you can double-click the split to reset it to its original position.

I intentionally did not tie the split position into app state, so the control will be more reusable in the future, if we move it to react common.

It'd be nice to merge this and the existing VerticalResizeContainer, but they've approached the problem from two somewhat different angles that makes the refactor a bit messy. The VerticalResizeContainer is a single window that can have its height changed, but it doesn't care what's shrinking/expanding below it. Meanwhile, this is a container for two different panels that manages both sides. In the future, I think it'd make sense to move this SplitPanel into React Common and take anything that's using the VerticalResizeContainer and have it use the SplitPanel instead, but I wanted to save that for another change.