Initial Parameterized Input Support (#9911)
This adds support for string and number parameters in criteria. I've tried to make the UI passable, but it's not perfect (particularly, I think the text boxes should expand to fit contents, it lacks signal when there's an issue, like an empty parameter, and the results page should probably put boxes around the parameter values). I'd like to follow up with a separate change to polish all that off, since this one is a bit big already.

One notable adjustment, I made the parameter path field into a list so that one parameter can affect multiple values in a validator plan (useful if the thing you're validating requires multiple checks, like custom functions are created and called).