Teacher Tool: Improve Customizable Criteria UI (#9915)
This change puts parameters into a clear error-state when they're empty, updates the results view to box in parameters per the design, and contains a few other minor tweaks throughout.

The boxes around parameter values in the results view seemed very similar to the boxes in the catalog view, so I refactored that into a shareable component. The main difference between the two is whether or not you show the parameter name vs the parameter value, which is indicated by whether or not you pass a criteria instance or just a catalog criteria.