Add catalog and validator plans for ai questions. (#9950) (#9951)
The implementation is not included here, but the docs changes are still needed to have the criteria appear in uploaded targets.

Of note: this does introduce the idea of a "system" parameter, which is an input to the validator plan which can change based on context, but which the user doesn't really need to specify in the crtitera template. In this case, that's the share id. The plan is for the teacher tool to automatically populate system parameters based on the provided key.