Teacher Tool: Fix Parameters on Criteria Instances that Reference the Same Validator Plans (#9967)
While simplifying the catalog, I noticed parameters were overwriting each other when we had multiple criteria instances that referenced the same validator plan. This was happening because all of them were pointing to the same plan object. To fix it, we can make a deep copy of the validator plan before we modify it.

I also updated the logDebug function to accept a data parameter to help keep our logs a little tidier. But the main change is in runEvaluateAsync.ts.