Teacher Tool: Require Auth (#10065)
This change makes it so that users must be signed in to use the teacher tool. If they are not signed in, the sign-in modal will open automatically.

A lot of the code here is based on code in multiplayer and skillmap. In effect, we're just adding a userProfile item to state, then we add an authClient to handle signing in and out, which sets the profile appropriately. The header bar has been changed to display the usual profile icon and dropdown.

We don't actually do anything with the profile yet, but it'd be nice to save checklists to the account at some point. Out of scope for this change, though.