Teacher Tool: Add Optional Feedback on Criteria (#10067)
This change adds an optional flag on catalog criteria which indicates that we should request feedback from the user on the tool's response. It's currently only set to true on the copilot criteria. When feedback is provided, it simply sends a tick event (collecting actual prompt and response data will require a more thorough privacy review).

There's one slightly odd note about this change: it felt weird to lock a user into their first feedback selection they make, so I've allowed the response to be changed or cleared, but we can't "take back" a tick event. I considered adding a delay before sending the event, but that seemed flaky, like there could always be scenarios where the delay wasn't long enough or the user closed the page before it fired, so instead, each feedback event includes the previous value as well as the new one. We'll have to do a bit of math in our queries to account for that, but it shouldn't be too difficult, something like actual helpful = count(helpful == true) - count(previousValue == true)

Because I wanted it for my own debugging here, I also added teacher tool support for the consoleticks flag (uses same infrastructure that the main webapp uses).