Teacher Tool: Relative URL for AI Eval (also remove copilot query param) (#10128)
This is important for ensuring we hit microbit.org instead of makecode.com for authenticated requests (where the auth cookie is associated with microbit.org). It also means we can reuse AuthClient.staticApiAsync(), which is nice.

I did not end up using relative URLs for fetching projects, as they would typically try to access staging instead of prod (depending on how the dev has configured pxt.cloud.DEV_BACKEND), but testing with staging projects isn't well supported at the moment. We could change this later, but for now, I think just using pxt.Cloud.apiRoot is simplest.

I also removed the copilot query param. It was helpful for debugging and demos, but with auth required, it no longer really works.