Teacher Tool: Handle Reload on ServiceWorkerRegistered Events (#10135)
With microsoft/pxt-microbit#5875, we disabled the automatic webapp reload that occurs when a new service worker is registered. We did that because the unexpected reload was causing microsoft/pxt-microbit#5824 and microsoft/pxt-microbit#5822.

Now, however, we have to handle that refresh ourselves. In this change, we listen for the event added in #10120 and refresh the iframe (via changing the src) when we get it.