Teacher Tool: Ensure Run-All Re-Runs Everything (#10141)
The user interaction flag is what determines if we should bypass the "run if unchanged / modified" check. We just need to pass it through.

I've pushed the toast display code back up to the callers of runSingleEvaluateAsync since we want different behavior in a run-all scenario vs a single run scenario that can't be totally decided via user interaction.