More Parameter Validation and Limitations (Length Limit for AI Prompts) (#10170)
This change started with just adding a limit to the AI prompt length, but as I was adding yet another field to the CriteriaParameter type definition (maxCharacters) which only really applies to a subset of parameter types, I decided it was time to tidy that up. Hence, a larger change.

Now, each parameter type has a corresponding type which can have its own fields, and a validate function switches on that type to perform different validations as applicable. Currently, I've added maxCharacters for strings (used for the AI prompt), and min and max for numbers (tested manually but not used; I mostly wanted a proof of concept around 2 different classes with 2 different types of validation).