Fix Gray Bar on Tutorial Resize (#10245)
In the tutorial view, we were using the same VerticalResizeContainer component regardless of whether the tutorial was at the top (where you can actually resize the container) or the side (where resizing was disabled), and when switching between the two modes, that container was retaining state. As a result, if the user switched from top-bar to side-bar and then back to top-bar, the resize container would immediately pop back to its former height, but it didn't signal that change to the rest of the elements, so things were not positioning themselves to account for it.

I toyed with a few ways to fix this, but ultimately, I think it's odd for us to keep the VerticalResizeContainer around at all when the tutorial is in the side-bar (non-resizable) view. Feels like asking for weird disabled-resize-related bugs in the future. So I've gone with an approach that fully swaps out the container depending on whether or not resizing is enabled. The downside to this is that we actually do want to remember the height the user set when we swap back, so I've added a new state var to track the last user-set height, and then we can set the initial height accordingly if we re-enter resize mode. This required pushing the shouldResize check into state as well.

Finally, since I now have two possible "tutorialWrapper" components, I changed the id to a className. I'm not sure if it's actually an antipattern to give both components the same id when it's guaranteed that only one will be present at any given time, but it seemed like it'd be safer to switch it, just in case.