Re-enable intellisense on android (#10246)
We had disabled this because it didn't work, but now we need it, so I hope it just works? The change that disabled it was #7099, which references bug microsoft/pxt-microbit#3170. Thankfully, things are clearly better now than they were then, so there's a good chance this is fine. Will still need some testing, though.