Switch to blockly keyboard experiment
Replaces (at least for now) the Blockly keyboard-navigation plugin.

Ultimately the experiment code might result in an update to
keyboard-navigation but we want to track the experiment and keep
MakeCode up-to-date.

There's no accessible API to to enable/disable so I've dropped that code
- all you can do is enable the experiment. I think this is for the best
  but needs discussion.

No toolbox integration for now - I've just disabled those parts of the
code.