Remove unnecessary tab stops from toolbox (#4)
- Instead of toolbox categories taking focus, use focus on the parent tree element alongside existing key handling. 
- Advanced categories are now always in the dom but hidden using CSS to fix toolbox-Blockly interactions.
- Simplify the toolbox handleKeyDown function so that key events are always handled after changes in the way the Blockly keyboard navigation plugin manages focus. The previous work unintentional broke keyboard navigation in the JavaScript/Monaco toolbox implementation.

We need to follow up with a change to use aria better here. We're going to do that in a separate PR but I'm concerned that removing the individual tab stops may have made things worse unless we add active descendant so will tackle separately along with a more general review of the tree.