Tweak focus outlines for toolbox/workspace
Workspace one is similar to the Blockly test environment.  I've gone with black
for both for now. I think they need to match in color but the toolbar is a blue
so we can't use that.
The left edge is still problematic.
The SVG scrollbars and the blocks are drawn over the outline which feels weird.
Worth further discussion but this is clearer for test use.