Order Themes in Theme Picker (#10398)
This allows us to specify a weight for each theme. Themes will be ordered with higher weights at the bottom. If a theme does not have a weight, it is put last. I've also moved when themes are loaded to before the page is shown, so there is no time where the default theme will be visible (unless something goes wrong).