Theme Fixes: Sound Effect Editor & Field Editor Header Bar (#10400)
The header bars for field editors are hard-coded to match the colors of the blocks (mostly, though I did have to fix this for the sound effect editor in this change). With this in mind, we should also hard-code the editor toggle color to ensure good contrast.

Also, the sound effect editor graphs had some elements that were hard to see in dark mode. This includes some changes that should make them more apparent.