Fix Editor Tour in French (#10411)
There were two issues happening causing the bug:

Firstly, some of the strings (location & highlight) should not be translated but were translated. They're used by the code to reference positions and elements in the DOM respectively, so when they're in french, they don't map to anything. I've gone through and removed the translations in crowdin.

Secondly, some of the properties had spaces before the semicolon, i.e: location : center, and that was being left in for the property name. For these, I think it's safe to trim the string, hence this PR.