Fix high contrast mode for the image editor.
As far as I can tell, you can't set !important from inside the typescript, so I used a var to set the preview color and set important in css. I also changed the top, bottom, and side bars to all have black backgrounds, which is different from the released version but I think more in-line with high contrast in general.