High Contrast Fixes (#10418)
* Fix high contrast mode for the image editor.

As far as I can tell, you can't set !important from inside the typescript, so I used a var to set the preview color and set important in css. I also changed the top, bottom, and side bars to all have black backgrounds, which is different from the released version but I think more in-line with high contrast in general.

* Update hc toolbox to also color icons

* Move toolbox less into override css and remove unnecessary rules.

* Fix multiplayer buttons in high contrast

Setting background instead of background-color cleared background-position and background-size properties which are needed to arrange the image. All we actually want to do here is set the color.

* Add borders around more image editor controls in high contrast

* Fix image editor done button hover