Remove unnecessary lf which was causing bugs (#10428)
Nothing in this string is actually supposed to be translated, but we have seen several occasions where the %property was translated incorrectly/accidentally, and that causes bugs. I think it's safe to just remove the lf altogether here.

Fixes microsoft/pxt-arcade#6289 (though I also updated the broken translation, which should also fix it)