Fix Card Icons in Tutorials (#10436)
Mostly, this fixes microsoft/pxt-arcade#6817. To get the id-card-looking icon in the tutorial view, we add a "card" class to the icon's css, which conflicted with the "card" css class we use for code-cards. I just added a bit of special casing to avoid the conflict.

I have also added the teaching-bubble-button class to the buttons in the teaching bubble, because I need to be able to access them specifically to fix microsoft/pxt-arcade#6807. However, the main chunk of that fix will happen in pxt-arcade with microsoft/pxt-arcade#6825