Save Skillmap Theme Settings (#10450)
This change saves the user's preference when setting a theme in skillmaps, and it fixes an error where I was double-parsing the json string and hitting exceptions when it was empty.

It's a bit of a shot in the dark, but I'm hoping this will fix microsoft/pxt-arcade#6808 by ensuring any skillmap-specific setting stays in-sync with the main editor setting. Hard to say for sure if that's what was happening though, since we don't have a concrete repro.