Port to Stable: Themes Fixes (#10451)
* Some More Assorted Themes Fixes (#10448)

* Fix function argument type display

* Fix toolbox separator line

* Change color of "Open in VS Code" button (and similarly, "Host multiplayer game") to match under-sim buttons.

* Fix pxt.json edit page

* Change textbox color to look less disabled

* Theme workspace search

* Fix foreground / background mismatch (target-background1 was using target-foreground2).

No bug filed on this since target1 and target2 tend to have similar foregrounds, but it was obvious when I was working on the theme builder and I forgot to fix it until now.

* Save Skillmap Theme Settings (#10450)

This change saves the user's preference when setting a theme in skillmaps, and it fixes an error where I was double-parsing the json string and hitting exceptions when it was empty.

It's a bit of a shot in the dark, but I'm hoping this will fix microsoft/pxt-arcade#6808 by ensuring any skillmap-specific setting stays in-sync with the main editor setting. Hard to say for sure if that's what was happening though, since we don't have a concrete repro.