Remove the loading state set when the iframe load event fires.
With the Read&Write extension enabled, the load event was getting delayed and we were actually setting state to "loading" after the newproject event fired, so then when the user tried to start an activity, the app still thought it was loading and failed.

I looked into why we were setting the loading state here (the load event typically fires when a page is finished loading, so it felt a little strange), and it was to fix a bug several years ago. It was a niche live -> beta scenario, and I'm unable to repro it now, so I've just removed the code.

Load docs: https://developer.mozilla.org/en-US/docs/Web/API/Window/load_event