Remove the loading state set when the iframe load event fires. (#10516)
Fixes microsoft/pxt-arcade#6893 (Hopefully. I have a repro, but it's not exact...)

With the "Read&Write for Google Chrome" extension enabled, the load event on the skillmap iframe gets delayed, and we were actually setting state to loading after the newproject event fired, so when the user tried to start an activity, the app still thought it was loading and failed.

I looked into why we were setting the loading state here. Turns out, it was to fix a bug several years ago (#8398). It was a niche live -> beta scenario, and I'm unable to repro it now, so I've just removed the code. The only way (that I know of) to trigger the load event on the iframe is to refresh the page, which also triggers recreation of the makecodeFrame, so the state was already set to loading in the constructor.