Fix show leds outline on Safari/Firefox.
Reinstate the blocklyFieldRect so it can get a rect-based outline because we
can't do a CSS outline on a group in Safari. That outline also had the wrong
size on Firefox for unclear reasons.

Needs some ugly CSS to deal with the hover effect from Blockly but that seems
an OK trade for the same DOM structure as other fields.

There's still an issue with the individual LEDs in Safari only.