Blockly 12 + keyboard nav plugin update (#10596)
* Upgrade to Blockly 12.0.1-beta.1 and plugin 0.0.8

Updates for:
- Awkward change in addConnectionHighlight that means Blockly now manages
  visibility and creates the highlights sooner.
- setEnabled removal
- dropdowns HTMLElement support
- plugin CSS
- toolbox integration with the plugin
- Reinstate outline around showLeds 'field' when focussed
- Active focus styling for flyout buttons and labels
- Fix focus handling for showLeds field editor
  The escape key now works as expected.
- Correct background color for play sound effect field
- We can now remove the patch script as isFullBlockField is public.
- Dropdown focus issue where dropdown div is used but widget div is focused

* Fix error clicking on comments

Explictly implement IBubble and add an implementation of the new-in-v12 method.

* Call Blockly.hideChaff when focusing the workspace

This hides the flyout if open fixing the global nav focusing the workspace when
the flyout is open.

* Upgrade to latest plugin release

* Fix melody field navigation

Previously there was a separate but invisible position for the field.
This field should be full block, like Blockly colour fields.

* Fix keyboard controls help for new plugin version

Reduce impact of missing/changed action names.

* Fix logged error in initCopyPaste

The keybindings are now available as an array due to https://github.com/google/blockly/pull/9
047
The Alt bindings were removed so an undefined entry was causing a conflict.

* Fix mouse and keyboard interaction with the showLeds block

Make sure that event listeners are added and removed at the right
points, and that fields are cleared properly.

Return ephemeral focus on pointer events as well as "Escape".

* Return ephemeral focus when the showLeds node is blurred

* CSS tweaks for Safari + image fields

* Fix more focus outlines.

* Fix show leds outline on Safari/Firefox.

Reinstate the blocklyFieldRect so it can get a rect-based outline because we
can't do a CSS outline on a group in Safari. That outline also had the wrong
size on Firefox for unclear reasons.

Needs some ugly CSS to deal with the hover effect from Blockly but that seems
an OK trade for the same DOM structure as other fields.

There's still an issue with the individual LEDs in Safari only.

* Fix missing passive focus styling.

* Fix passive focus case that also needs to be more specific.

* Correct disabled reason to use new constant

* Fix addConnectionHighlight issues

This becomes more necessary in plugin versions where Blockly have restored
navigating to empty value connections but is an improvement anyway.

Behaviour change: the connection indicators now work for and/or blocks with
empty value connections.

* Fix closing chaff when clicking on toolbox

---------

Co-authored-by: Robert Knight <robert@microbit.org>