Upgrade to Blockly 12.1.0-beta.0 (#10616)
* Update to latest beta version of Blockly and keyboard plugin

* Updates to integrate with Blockly focus manager changes

* Minor changes for Monaco flyout

* Fix tabbing forwards from the toolbox

If keyboard controls are off, tab to the workspace. Otherwise,
move the cursor to flyout taking into account the caching
flyout implementation.

* Set keyboardNavigationController to true when key controls on

This Blockly setting enables focus indicators around the toolbox,
flyout and workspace, but doesn't automatically turn on if the
user tabs to the workspace.

* Remove tab stops in Monaco flyout

Use roving tab index instead. Align focus-visible styles of labels
with blocks.

* New fixes for field_ledmatrix

* Tidy imports