Add Evaluate Button to Share Page (#10614)
This adds an "Evaluate" button to the share page, which will only appear if the teachertool/showSharePageEvalButton is set to true in targetconfig.

It also includes some fixes to the load-project-from-url behavior that already existed in the teacher tool. Specifically, it:

1. Ensures we preserve the hash and params on sign-in.
2. Updates the share link input.
3. Loads a new default checklist and goes to the results page automatically

The default checklist will only load if the teacher does not already have a checklist loaded. This way, if they're grading the same project for a bunch of students and have a customized rubric, they can still click the evaluate button here and it will load the same rubric they've been working with. It also prevents any accidental loss of work, which could happen if we overwrote the active rubric.