Add support for region-based feature flags (#10621)
This adds support for feature flags that can be enabled (or disabled) for specific regions, based on the backend changes here: microsoft/pxt-backend#1117

I've created a more generic "enabledFeatures" section of the AppTheme in pxtarget in an attempt to manage the proliferation of top-level feature switches. If no includeRegions or excludeRegions are specified, then the feature is enabled globally. Otherwise, it will only be enabled if the user's region matches one of the includeRegions and does not match any of the excludeRegions.

For development purposes, I've added a DEV_REGION constant that we can change similar to how we manage DEV_BACKEND, which allows us to fake being in a different region on local builds. (This also simplified things, since our localhost cdn logic is a little weird and didn't actually go to staging.) In my own testing, however, I did ensure it works with an actual staging CDN URL, too.