Rephrase Error Help Button to Indicate AI Usage (#10645)
There was some concern over the fact that the "Help me understand" button did not clearly indicate that it would send your code to AI. To address this, I've changed the wording on the button. I've also changed the icon the sparkle, which seems to be becoming a standard AI-indicator.