Poll for error help response (#10659)
This switches the frontend code so it polls the backend for a response to its error-help request. It introduces a new, simplistic exponential backoff utility function to help, so we don't spam ourselves (and deep prompt) with too many status requests.

Separately, this also fixes sets an error region when we can't detect the actual region to avoid repeated errors getting thrown for failing to initialize the region.