Make initRegionAsync more resilient (#10670)
This fixes the static package scenario by making initRegionAsync more resilient to network & URL format failures. I couldn't find a reliable way to detect if we were in "static package" mode (and adding one seemed out of scope for this fix). But I think it can be fixed more simply: instead of throwing an exception when we can't get the region, just log the error to console, set a default, and move on. There's no particularly good reason to fail noisily in this case, so I think the simple log is sufficient.