Move "Debug My Code Button" to problems header & show in blocks (#10669)
This moves the "Debug My Code" button into the problem header, and it adds the button for blocks as well as text.

This button does not make sense when you have a compile error, since you can't debug a program that can't run. As such, the existing behavior checks for a stack trace before showing the button. I've preserved this in the monaco scenario, but in blocks, it's rare to get a compile error that actually stops the code from running, so it simply always shows the debug button for blocks (some of our errors do lack stack traces, otherwise it'd be a non-issue).