Keyboard Controls help tweaks (#10703)
* Keyboard navigation help tweaks

- Clarify move mode in help docs
- Avoid term region in favour of area based on advice on comprehension
- Allow Ctrl/Cmd+B to toggle instead of just open the area menu based on user feedback

* Prefer simpler shortcuts to get preferred redo shortcut on Windows.

I've also opened a Blockly PR to reorder the Shift+Cmd+Z shortcut to the more
conventional Cmd+Shift+Z.

Tweak platform detection to make it easier to check Windows locally.

* Fix restoring focus when toggling area menu via Ctrl+B

* Prevent dropdown div showing over toolbox/area 3

I originally tried to hideChaff when opening the overlay but that seems to move
the selection in a delayed way that breaks focus on the overlay.

---------

Co-authored-by: Richard Knoll <riknoll@users.noreply.github.com>