Detect the on start block in stack traces by checking for the root function (#10715)
The JS code for anything in "on start" is just at the root level of the javascript, which leads to some confusion when we parse the locInfo associated with an "on start" stack frame.

In these cases, we can detect we're at the root level (therefore in "on start") by checking if the function name is the root function, "<main>".